Criação dos cards empresas parceiras
