--- src/components/Header/header.jsx
@@ -1,19 +1,0 @@
-import "./style.css";
-
-function Header() {
-  return (
-    <header className="header">
-      <div className="headerBackground">
-        <div className="headerInfos">
-          <span>Inicio</span>
-          <span> O que é selo verde </span>
-          <span>Pontuação</span>
-          <span>Vouchers</span>
-          <span>Empresas parceiras</span>
-        </div>
-      </div>
-    </header>
-  );
-}
-
-export default Header;
