+++ src/components/Header/index.jsx
@@ -1,0 +1,19 @@
+import "./style.css";
+
+function Header() {
+  return (
+    <header className="header">
+      <div className="headerBackground">
+        <div className="headerInfos">
+          <span>Inicio</span>
+          <span> O que é selo verde </span>
+          <span>Pontuação</span>
+          <span>Vouchers</span>
+          <span>Empresas parceiras</span>
+        </div>
+      </div>
+    </header>
+  );
+}
+
+export default Header;
