+++ src/pages/UserProfile/index.jsx
@@ -1,6 +1,6 @@
 // import Button from "../../components/Button";
 // import { SolidSmallButton } from "../../components/Button/styles";
-import Header from "../../components/Header/Header";
+import Header from "../../components/Header/header";
 import Footer from "../../components/Footer";
 import Main from "../../components/Main";
 import { StyledImage, StyledSection } from "./styles";
